pyproject.toml
[tool.poetry]
name = "charlie-backend"
version = "0.1.0"
description = "Charlie AI Assistant Backend"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
supabase = "^2.0.0"
google-cloud-speech = "^2.22.0"
google-cloud-texttospeech = "^2.16.0"
google-generativeai = "^0.3.0"
pyaudio = "^0.2.11"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.1.0"
celery = "^5.3.0"
redis = "^5.0.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiofiles = "^23.2.1"
librosa = "^0.10.1"
soundfile = "^0.12.1"
asyncio-mqtt = "^0.13.0"
psutil = "^5.9.6"
slowapi = "^0.1.9"
pillow = "^11.2.1"
pywin32 = {version = "^306", markers = "sys_platform == 'win32'"}
google-auth = "^2.16.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.71.0"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true 